requirements.txt
flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
pycryptodome==3.19.0
cryptography==41.0.3
gym==0.26.2
stable-baselines3==2.1.0
nistrng==1.0.0
joblib==1.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3 